Fixed bugs in mobile layout
The majority of bugs was in the deletion functions in the different
modes. Also a new words can now be typed in the letter selection mode
where they are successfully added to the dictionary in lowercase.
Previously a word added would cause a nullpointer when accesed from the
dictionary.  Some other small fixes too.