Dictionary bugfix and ListView scroll improvements
The dictionary on mobile is now working as expected. The dictionary
lists on mobile, adaptive and etos now has a improved scroll feature
which scrolls the selected element to the top when it exceeds or equals
the index of the visible list midpoint. Cleaned up unnecesary log
statements in mobile layout.